Plugin for upcoming feature LUCENE-3071
In contrario to the pathhierarchy-update branch, this implementation (with the same functionnality) takes the form of a plugin.
The advantage of making a plugin, is that it is easily distribuable and pluggable.
The drawback is that the registered tokenizer name is suffixed by "_upcoming", not to conflict with the official implementation.
But such difference normally only impacts the file config/elasticsearch.yml.

(Following is the same commit log as pathhierarchy-update)
Skip and reverse parameters for [Reverse]PathHierarchyTokenizer(Upcoming).
Implemented as side classes not to conflict with existing analysis.path.PathHierarchyTokenizer from contrib.
Once the LUCENE-3071 feature will be available in Lucene 3.2 (maybe), [Reverse]PathHierarchyTokenizerUpcoming classes should be removed, and PathHierarchyTokenizerFactory should be updated accordingly.